test: extract long-running e2e tests to new Jest project, run it and wallet tests in nightly
