+++ packages/e2e/test/long-running/delegation-rewards.test.ts
@@ -1,11 +1,11 @@
 import * as envalid from 'envalid';
 import { Cardano, createSlotEpochCalc } from '@cardano-sdk/core';
 import { SignedTx, SingleAddressWallet, buildTx } from '@cardano-sdk/wallet';
-import { TestWallet, getWallet } from '../../../src';
-import { assertTxIsValid, waitForWalletStateSettle } from '../../../../wallet/test/util';
+import { TestWallet, getWallet } from '../../src';
+import { assertTxIsValid, waitForWalletStateSettle } from '../../../wallet/test/util';
 import { filter, firstValueFrom } from 'rxjs';
 import { logger } from '@cardano-sdk/util-dev';
-import { requestCoins, runningAgainstLocalNetwork, submitAndConfirm, transferCoins, waitForEpoch } from '../../util';
+import { requestCoins, runningAgainstLocalNetwork, submitAndConfirm, transferCoins, waitForEpoch } from '../util';
 
 // Verify environment.
 export const env = envalid.cleanEnv(process.env, {
